# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: WashBuddy API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /providers/{providerId}/locations/{locationId}/services:
    get:
      operationId: listServices
      tags: [services]
      summary: List services for a location
      parameters:
        - $ref: "#/components/parameters/ProviderId"
        - $ref: "#/components/parameters/LocationId"
      responses:
        "200":
          description: Service list
          content:
            application/json:
              schema:
                type: object
                properties:
                  services:
                    type: array
                    items:
                      $ref: "#/components/schemas/Service"
                required: [services]
    post:
      operationId: createService
      tags: [services]
      summary: Create service
      parameters:
        - $ref: "#/components/parameters/ProviderId"
        - $ref: "#/components/parameters/LocationId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateServiceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  service:
                    $ref: "#/components/schemas/Service"
                required: [service]
  /providers/{providerId}/locations/{locationId}/services/{serviceId}:
    patch:
      operationId: updateService
      tags: [services]
      summary: Update service
      parameters:
        - $ref: "#/components/parameters/ProviderId"
        - $ref: "#/components/parameters/LocationId"
        - $ref: "#/components/parameters/ServiceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateServiceInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  service:
                    $ref: "#/components/schemas/Service"
                required: [service]
  /vehicles:
    get:
      operationId: listVehicles
      tags: [vehicles]
      summary: List vehicles
      responses:
        "200":
          description: Vehicle list
          content:
            application/json:
              schema:
                type: object
                properties:
                  vehicles:
                    type: array
                    items:
                      $ref: "#/components/schemas/Vehicle"
                required: [vehicles]
    post:
      operationId: createVehicle
      tags: [vehicles]
      summary: Create vehicle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVehicleInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  vehicle:
                    $ref: "#/components/schemas/Vehicle"
                required: [vehicle]
  /vehicles/{vehicleId}:
    patch:
      operationId: updateVehicle
      tags: [vehicles]
      summary: Update vehicle
      parameters:
        - name: vehicleId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateVehicleInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  vehicle:
                    $ref: "#/components/schemas/Vehicle"
                required: [vehicle]
...
components:
  parameters:
    ProviderId:
      name: providerId
      in: path
      required: true
      schema:
        type: string
    LocationId:
      name: locationId
      in: path
      required: true
      schema:
        type: string
    ServiceId:
      name: serviceId
      in: path
      required: true
      schema:
        type: string
...
  schemas:
...
    Service:
      type: object
      properties:
        id:
          type: string
        locationId:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        durationMins:
          type: integer
        basePriceMinor:
          type: integer
        platformFeeMinor:
          type: integer
        currencyCode:
          type: string
        capacityPerSlot:
          type: integer
        leadTimeMins:
          type: integer
        requiresConfirmation:
          type: boolean
        isVisible:
          type: boolean
      required: [id, locationId, name, durationMins, basePriceMinor, platformFeeMinor, currencyCode, capacityPerSlot, isVisible]
...
    CreateServiceInput:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        durationMins:
          type: integer
        basePriceMinor:
          type: integer
        platformFeeMinor:
          type: integer
        capacityPerSlot:
          type: integer
        leadTimeMins:
          type: integer
        requiresConfirmation:
          type: boolean
      required: [name, durationMins, basePriceMinor]
    UpdateServiceInput:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        durationMins:
          type: integer
        basePriceMinor:
          type: integer
        platformFeeMinor:
          type: integer
        capacityPerSlot:
          type: integer
        leadTimeMins:
          type: integer
        isVisible:
          type: boolean
    Vehicle:
      type: object
      properties:
        id:
          type: string
        ownerId:
          type: string
        fleetId:
          type: ["string", "null"]
        unitNumber:
          type: string
        vin:
          type: ["string", "null"]
        licensePlate:
          type: ["string", "null"]
        categoryCode:
          type: string
        subtypeCode:
          type: ["string", "null"]
        lengthFt:
          type: ["number", "null"]
        heightFt:
          type: ["number", "null"]
        isActive:
          type: boolean
      required: [id, ownerId, unitNumber, categoryCode, isActive]
    CreateVehicleInput:
      type: object
      properties:
        unitNumber:
          type: string
        vin:
          type: string
        licensePlate:
          type: string
        categoryCode:
          type: string
        subtypeCode:
          type: string
        lengthFt:
          type: number
        heightFt:
          type: number
        fleetId:
          type: string
      required: [unitNumber, categoryCode]
    UpdateVehicleInput:
      type: object
      properties:
        unitNumber:
          type: string
        licensePlate:
          type: string
        categoryCode:
          type: string
        subtypeCode:
          type: string
        lengthFt:
          type: number
        heightFt:
          type: number
        isActive:
          type: boolean
...